+++ RELEASES.md
@@ -735,3 +735,8 @@
   - refactor: prioritization of data providers by length of filter pattern desc
   - fix: when using volatile permissions, system permissions starting with _ are retained in same datastore as groups
   - fix #283: client connection status is set to ACTIVE before authentication happens, added new status RECONNECT_ACTIVE
+
+11.2.5 2020-03-27
+-----------------
+  - test: per message deflate compression configuration
+  - doc: per message deflate compression README update